+++ starters/backstage/src/config/index.ts
@@ -1,0 +1,1 @@
+export {BackstageConfiguration} from "./BackstageConfiguration";