+++ starters/backstage/src/proxy/index.ts
@@ -1,0 +1,1 @@
+export {CatalogClientProxy} from "./CatalogClientProxy";