+++ samples/sample-express-mongodb/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @nodeboot/express-sample-mongodb
 
+## 2.0.8
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/core@1.5.3
+    -   @nodeboot/express-server@1.2.7
+    -   @nodeboot/starter-firebase@1.0.10
+    -   @nodeboot/starter-persistence@1.6.3
+    -   @nodeboot/starter-validation@1.0.5
+
 ## 2.0.7
 
 ### Patch Changes