+++ tests/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nodeboot/test
 
+## 1.1.10
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/core@1.5.3
+
 ## 1.1.9
 
 ### Patch Changes